Moved glossary to a task
